crates `svm-runtime` and `svm-runtime-c-api` have no missing docs or unreachable code.
