+++ crates/svm-runtime/src/import.rs
@@ -1,4 +1,4 @@
-use std::convert::{TryFrom, TryInto};
+use std::convert::TryFrom;
 use std::ffi::c_void;
 use std::sync::Arc;
 
@@ -6,7 +6,6 @@ use crate::Context;
 
 use wasmer::{
     Export, Exportable, Function, FunctionType, RuntimeError, Store, Type as WasmerType, Val,
-    WasmerEnv,
 };
 
 use svm_ffi::{svm_byte_array, svm_env_t, svm_func_callback_t};
@@ -14,6 +13,8 @@ use svm_types::{Type, WasmType, WasmValue};
 
 static WASMER_ARGS_STR: Type = Type::Str("Wasmer Args");
 
+/// Holds data about an import of a function
+/// conforming to the FFI interface
 #[derive(Debug, Clone)]
 pub struct ExternImport {
     name: String,
@@ -30,6 +31,7 @@ pub struct ExternImport {
 }
 
 impl ExternImport {
+    /// Creates a new struct
     pub fn new(
         name: String,
         namespace: String,
@@ -48,6 +50,7 @@ impl ExternImport {
         }
     }
 
+    /// Wraps as a `Wasmer Export`
     pub fn wasmer_export(&self, store: &Store, ctx: &mut Context) -> (Export, *mut svm_env_t) {
         unsafe {
             // The following code has been highly influenced by code here:
@@ -108,25 +111,25 @@ impl ExternImport {
 
             let func_ty = self.wasmer_function_ty();
 
-            /// making the input `&mut Context` appear as `*const c_void`
+            // making the input `&mut Context` appear as `*const c_void`
             let inner_env = ctx as *mut Context as *const c_void;
             let host_env = self.host_env;
 
-            /// The import used `env` (using Wasmer terminology) will be a struct of `svm_env_t`
-            /// This `#[repr(C)]` struct will contain two pointers to two types of `env`:
-            ///
-            /// 1. SVM inner env - a pointer to the `Context`
-            ///    Once SVM has finished executing a transaction its memory will be deallocated.
-            ///
-            /// 2. Host env - a pointer given as input by the so-called `Host`
-            ///    The responsibility of release that memory is up to the caller (the `host`).
+            // The import used `env` (using Wasmer terminology) will be a struct of `svm_env_t`
+            // This `#[repr(C)]` struct will contain two pointers to two types of `env`:
+            //
+            // 1. SVM inner env - a pointer to the `Context`
+            //    Once SVM has finished executing a transaction its memory will be deallocated.
+            //
+            // 2. Host env - a pointer given as input by the so-called `Host`
+            //    The responsibility of release that memory is up to the caller (the `host`).
             let func_env = svm_env_t {
                 inner_env,
                 host_env,
             };
 
-            /// The heap-allocated `func_env` will be deallocated by later by `SVM` running runtime.
-            /// (See method `funcs_envs_destroy` under `src/runtime/default.rs`)
+            // The heap-allocated `func_env` will be deallocated by later by `SVM` running runtime.
+            // (See method `funcs_envs_destroy` under `src/runtime/default.rs`)
             let ty = Type::of::<svm_env_t>();
             let func_env = svm_ffi::into_raw(ty, func_env) as *mut svm_env_t;
 
@@ -138,10 +141,12 @@ impl ExternImport {
         }
     }
 
+    /// Returns the import's name
     pub fn name(&self) -> &str {
         &self.name
     }
 
+    /// Returns the import's namespace
     pub fn namespace(&self) -> &str {
         &self.namespace
     }
@@ -160,7 +165,6 @@ fn to_wasmer_types(types: &[WasmType]) -> Vec<WasmerType> {
         .map(|ty| match ty {
             WasmType::I32 => WasmerType::I32,
             WasmType::I64 => WasmerType::I64,
-            _ => panic!("Only i32 and i64 are supported."),
         })
         .collect()
 }
