+++ crates/svm-runtime/src/lib.rs
@@ -1,7 +1,7 @@
-#![allow(missing_docs)]
-#![allow(unused)]
-#![allow(dead_code)]
-#![allow(unreachable_code)]
+#![deny(missing_docs)]
+#![deny(unused)]
+#![deny(dead_code)]
+#![deny(unreachable_code)]
 #![feature(vec_into_raw_parts)]
 
 //! `SVM-runtime` crate is the glue between `SVM` to a Wasm Runtime
@@ -19,17 +19,19 @@ pub use runtime::create_rocksdb_runtime;
 /// Gas estimation and metering.
 pub mod gas;
 
+/// Managing the runtime's environment
+pub mod env;
+
 mod import;
 mod storage;
 
-pub mod env;
 pub use import::ExternImport;
 
 /// Implements `Context`. Used for managing data of running `SVM` apps.
 mod context;
 pub use context::Context;
 
-/// Implements common functionalities to be consnumed by tests.
+/// Implements common functionality to be consnumed by tests.
 pub mod testing;
 
 /// Implements the `SVM` vmcalls (a.k.a libcalls / hostcalls / syscalls)
