+++ crates/svm-runtime/src/vmcalls/mod.rs
@@ -1,4 +1,4 @@
-use wasmer::{Exports, Function, ImportObject, Store};
+use wasmer::{Exports, Function, Store};
 
 use crate::Context;
 
@@ -21,6 +21,8 @@ macro_rules! func {
     }};
 }
 
+/// Registers SVM internal host functions (a.k.a vmacalls) into
+/// Into `Wasmer` Import Object (it's done by inserting to input `Exports`)
 pub fn wasmer_register(store: &Store, ctx: &Context, ns: &mut Exports) {
     ns.insert("svm_calldata_offset", func!(store, ctx, calldata_offset));
     ns.insert("svm_calldata_len", func!(store, ctx, calldata_len));
